# assets/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "JoustLM API",
        "description": "API for LLM-powered knowledge extraction and management",
        "version": "1.0.0",
        "contact": {
            "name": "JoustLM Team"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/v1/auth/signup": {
            "post": {
                "summary": "User signup",
                "description": "Create a new user account",
                "tags": [
                    "Authentication"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignupRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SignupResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Username already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "summary": "User login",
                "description": "Authenticate user and get access token",
                "tags": [
                    "Authentication"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "summary": "User logout",
                "description": "Logout user and invalidate session",
                "tags": [
                    "Authentication"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/extract": {
            "post": {
                "summary": "Extract knowledge from text",
                "description": "Analyze text using LLM to extract knowledge",
                "tags": [
                    "LLM Analysis"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LLMAnalysisRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Analysis completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LLMAnalysisResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/knowledge": {
            "get": {
                "summary": "Get knowledge entries",
                "description": "Retrieve paginated knowledge entries for the authenticated user",
                "tags": [
                    "Knowledge Base"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100,
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Knowledge entries retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetKnowledgeResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create knowledge entry",
                "description": "Create a new knowledge entry",
                "tags": [
                    "Knowledge Base"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateKnowledgeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Knowledge entry created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/KnowledgeBaseEntry"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/search": {
            "get": {
                "summary": "Search knowledge base",
                "description": "Search through knowledge entries",
                "tags": [
                    "Search"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "topic",
                        "in": "query",
                        "description": "Filter by topic",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Filter by keyword",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sentiment",
                        "in": "query",
                        "description": "Filter by sentiment",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "positive",
                                "negative",
                                "neutral"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100,
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SearchResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
...
            "SignupRequest": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string",
                        "description": "Username for the new account",
                        "example": "john_doe"
                    },
                    "password": {
                        "type": "string",
                        "minLength": 6,
                        "description": "Password for the new account",
                        "example": "securepass123"
                    }
                },
                "required": [
                    "username",
                    "password"
                ],
                "example": {
                    "username": "john_doe",
                    "password": "securepass123"
                }
            },
            "SignupResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string",
                        "description": "JWT access token",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                    },
                    "refresh_token": {
                        "type": "string",
                        "description": "Refresh token",
                        "example": "dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4gZXhhbXBsZQ=="
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Token expiration time",
                        "example": "2024-01-15T10:30:00Z"
                    }
                },
                "required": [
                    "token",
                    "refresh_token",
                    "expires_at"
                ],
                "example": {
                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
                    "refresh_token": "dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4gZXhhbXBsZQ==",
                    "expires_at": "2024-01-15T10:30:00Z"
                }
            },
            "LoginRequest": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string",
                        "description": "Username",
                        "example": "john_doe"
                    },
                    "password": {
                        "type": "string",
                        "description": "Password",
                        "example": "securepass123"
                    }
                },
                "required": [
                    "username",
                    "password"
                ],
                "example": {
                    "username": "john_doe",
                    "password": "securepass123"
                }
            },
            "LoginResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string",
                        "description": "JWT access token",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                    },
                    "refresh_token": {
                        "type": "string",
                        "description": "Refresh token",
                        "example": "dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4gZXhhbXBsZQ=="
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Token expiration time",
                        "example": "2024-01-15T10:30:00Z"
                    }
                },
                "required": [
                    "token",
                    "refresh_token",
                    "expires_at"
                ],
                "example": {
                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
                    "refresh_token": "dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4gZXhhbXBsZQ==",
                    "expires_at": "2024-01-15T10:30:00Z"
                }
            },
            "LLMAnalysisRequest": {
                "type": "object",
                "properties": {
                    "text": {
                        "type": "string",
                        "description": "Text to analyze",
                        "example": "Artificial intelligence is transforming the way we work and live. From machine learning algorithms that power recommendation systems to natural language processing that enables chatbots, AI technologies are becoming increasingly sophisticated and integrated into our daily lives."
                    },
                    "model": {
                        "type": "string",
                        "description": "LLM model to use for analysis",
                        "example": "gemini-2.0-flash"
                    }
                },
                "required": [
                    "text",
                    "model"
                ],
                "example": {
                    "text": "Artificial intelligence is transforming the way we work and live. From machine learning algorithms that power recommendation systems to natural language processing that enables chatbots, AI technologies are becoming increasingly sophisticated and integrated into our daily lives.",
                    "model": "gemini-2.0-flash"
                }
            },
            "LLMAnalysisResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Unique analysis ID",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "text": {
                        "type": "string",
                        "description": "Original text analyzed",
                        "example": "Artificial intelligence is transforming the way we work and live. From machine learning algorithms that power recommendation systems to natural language processing that enables chatbots, AI technologies are becoming increasingly sophisticated and integrated into our daily lives."
                    },
                    "title": {
                        "type": "string",
                        "description": "Generated title for the analysis",
                        "example": "The Impact of Artificial Intelligence on Modern Society"
                    },
                    "summary": {
                        "type": "string",
                        "description": "Summary of the text",
                        "example": "This text discusses how artificial intelligence is revolutionizing various aspects of human life, including work environments and daily activities. It highlights key AI technologies like machine learning and natural language processing, emphasizing their growing sophistication and integration into everyday experiences."
                    },
                    "topics": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "Identified topics",
                        "example": [
                            "artificial intelligence",
                            "machine learning",
                            "technology",
                            "automation",
                            "natural language processing"
                        ]
                    },
                    "sentiment": {
                        "type": "string",
                        "enum": [
                            "positive",
                            "negative",
                            "neutral"
                        ],
                        "description": "Sentiment analysis result",
                        "example": "positive"
                    },
                    "keywords": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "Extracted keywords",
                        "example": [
                            "AI",
                            "machine learning",
                            "algorithms",
                            "recommendation systems",
                            "chatbots",
                            "NLP",
                            "technologies"
                        ]
                    },
                    "confidence": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "maximum": 1,
                        "description": "Confidence score of the analysis",
                        "example": 0.87
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Analysis creation timestamp",
                        "example": "2024-01-15T10:30:00Z"
                    },
                    "error": {
                        "type": "string",
                        "description": "Error message if analysis failed",
                        "example": null
                    }
                },
                "required": [
                    "id",
                    "text",
                    "summary",
                    "topics",
                    "sentiment",
                    "keywords",
                    "confidence",
                    "created_at"
                ],
                "example": {
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "text": "Artificial intelligence is transforming the way we work and live. From machine learning algorithms that power recommendation systems to natural language processing that enables chatbots, AI technologies are becoming increasingly sophisticated and integrated into our daily lives.",
                    "title": "The Impact of Artificial Intelligence on Modern Society",
                    "summary": "This text discusses how artificial intelligence is revolutionizing various aspects of human life, including work environments and daily activities. It highlights key AI technologies like machine learning and natural language processing, emphasizing their growing sophistication and integration into everyday experiences.",
                    "topics": [
                        "artificial intelligence",
                        "machine learning",
                        "technology",
                        "automation",
                        "natural language processing"
                    ],
                    "sentiment": "positive",
                    "keywords": [
                        "AI",
                        "machine learning",
                        "algorithms",
                        "recommendation systems",
                        "chatbots",
                        "NLP",
                        "technologies"
                    ],
                    "confidence": 0.87,
                    "created_at": "2024-01-15T10:30:00Z",
                    "error": null
                }
            },
...
            "GetKnowledgeResponse": {
                "type": "object",
                "properties": {
                    "knowledge": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/KnowledgeBaseEntry"
                        },
                        "description": "List of knowledge entries"
                    },
                    "currentPage": {
                        "type": "integer",
                        "description": "Current page number"
                    },
                    "totalPages": {
                        "type": "integer",
                        "description": "Total number of pages"
                    },
                    "totalCount": {
                        "type": "integer",
                        "description": "Total number of entries"
                    }
                },
                "required": [
                    "knowledge",
                    "currentPage",
                    "totalPages",
                    "totalCount"
                ]
            },
...
            "CommonResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "Operation success status",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "description": "Response message",
                        "example": "Operation completed successfully"
                    },
                    "data": {
                        "type": "object",
                        "description": "Response data",
                        "example": {}
                    }
                },
                "required": [
                    "success"
                ],
                "example": {
                    "success": true,
                    "message": "Operation completed successfully",
                    "data": {}
                }
            },
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "description": "HTTP status code",
                        "example": 400
                    },
                    "message": {
                        "type": "string",
                        "description": "Error message",
                        "example": "Invalid request parameters"
                    }
                },
                "required": [
                    "code",
                    "message"
                ],
                "example": {
                    "code": 400,
                    "message": "Invalid request parameters"
                }
            }
        }
    },
...
}